settings: simplify stream key handling
The specification of valid stream keys in settings.json was
previously defined for each network. During early development
there had been differences between quality settings and the
associated stream keys. These differences could be now unified
and thus it isn't needed anymore to specify the mapping between
quality_key and stream_key for each network separately.

The implementation in settings.py has been modified with the goal
to keep other parts of the code untouched. Therefore the streams
definition is now given to the network class.

This also leaves the possibility to react on future changes of
the mapping between quality_key and stream_key in the AudioAddict
API to go back to a more differentiated specification if needed
without touching too much code.