settings: avoid redundant key names
In settings.json network_key names had been maintained twice.
One time for each network as the dictionary key and a second
time in the network_order list.

Ordering of the networks is now maintained via a sort_key attribute
of each network.

The code in settings.py was adapted to this change.