play: add new way of listen_key handling
The listen_key gets now retrieved via the 'auth' module.

The listen_key gets validated via the query to the playlist API. If
the key is invalid than this could mean that the user was resetting
his/her listen_key and we can try to fetch it once more.

In case of a static configured listen_key an invalid listen_key will
immediately raise an error.