api: introduce User class
In order to wrap the user data returned by the autentication API,
the class User gets introduced. This class wraps the JSON response
and provides via properties the keys which are currently relevant
for further usage in the whole addon.

The User class shall also make clear that the returned data of the
authentication API is user related information. This class doesn't
reveal all of the user information available but only those which
are currently needed.