fix Yatse ListItem issue
If a channel was started via Yatse, Kodi wasn't showing the channel
image anymore. That was due to using xbmcplugin.setResolvedUrl()
which takes a ListItem object to show and this one was incomplete.

In Kodi context this isn't a problem since setResolvedUrl() will
update the currently selected ListItem but if started via Yatse,
there is no such item.

There is still an issue that Yatse can't show the channel image
itself. So far I haven't figured out why this is happening. This
issue will get fixed later.