change order of functions/methods
Functions should appear in top down order to avoid jumping in the
file during reading the code. This also makes it obvious for the
reader where to search a specific function definition in the code.

Example:
  def a():
    b()
    c()

  def b():
    pass

  def c():
    pass

This is a very drastic change which makes merges based on older
revisions more complicated but from my point of view necessary to
maintain a clean code structure.

Should have been this way from the beginning but from now on it
will be kept this way for new functions/methods to stay consistent.