add sorting by channel creation date
Channels can be sorted now by their creation date. This makes it
possible to sort channels in descending order to see which channels
have been added recently to a specific network.

For sorting the sort method "dateadded" is used. Unfortunately this
method belongs to the 'video' category but works as well for music
addons.

This implementation will stay until something more suitable might
be found.