api: better detection of disabled channels
There are channels which are marked differently in the channel list
as disabled as the majority. 99% of the hannels which are disabled
start with a capital 'X' as the channel name but there are also
channels which don't follow this convention. These channels have
nearly no meta data set.

For now I will make a weak check to detect these channels and just
look if there is any channel art. If there is no channel art
defined it is assumed that the channel is disabled.

The better way would be to check for those channels where this weak
check is true additionally if the playlist is empty. Currently this
would be hacky to implement so I leave it for later after the next
refactoring.