drop support for public streams
Sad but true...
Recently the icecast servers of the audioaddict network providing
the public streams have been disabled. How public streams are
nowadays provided isn't known to me and I'm not going to reverse
engineer it. This means I will drop public streams support and
concentrate from now on only on premium support which still works
as it did all the time.