api: improve supported channel detection
Supported channels are now detected via querying the ListenApi. It
returns a brief list of channels which are currently supported or
respectively active for a given network. The verbose channel list
is filtered according to the result of the ListenApi.

This simplifies the channel handling and makes it more reliable.